feat: Move sound peripherals into its own menu "Peripherals/Sound"
Created separate menu for sound peripherals, moved 4 devices there.
Moved sound related peripherals' source files into standalone folder (like disk related stuff is in)

See more
https://sourceforge.net/u/arki55/tickets/4/
