+++ fuse.c
@@ -65,12 +65,13 @@
 #include "module.h"
 #include "movie.h"
 #include "mempool.h"
-#include "peripherals/ay.h"
 #include "peripherals/dck.h"
+#include "peripherals/sound/ay.h"
+#include "peripherals/sound/fuller.h"
+#include "peripherals/sound/melodik.h"
 #include "peripherals/disk/beta.h"
 #include "peripherals/disk/didaktik.h"
 #include "peripherals/disk/fdd.h"
-#include "peripherals/fuller.h"
 #include "peripherals/ide/divide.h"
 #include "peripherals/ide/divmmc.h"
 #include "peripherals/ide/simpleide.h"
@@ -81,7 +82,6 @@
 #include "peripherals/if1.h"
 #include "peripherals/if2.h"
 #include "peripherals/kempmouse.h"
-#include "peripherals/melodik.h"
 #include "peripherals/multiface.h"
 #include "peripherals/printer.h"
 #include "peripherals/scld.h"
