+++ machine.h
@@ -29,9 +29,9 @@
 #include "libspectrum.h"
 
 #include "display.h"
-#include "peripherals/ay.h"
-#include "peripherals/covox.h"
-#include "peripherals/specdrum.h"
+#include "peripherals/sound/ay.h"
+#include "peripherals/sound/covox.h"
+#include "peripherals/sound/specdrum.h"
 #include "spectrum.h"
 
 typedef libspectrum_byte (*spectrum_unattached_port_fn)( void );
