+++ menu_data.dat
@@ -81,6 +81,7 @@ Options/_Sound..., Item
 
 Options/_Peripherals, Branch
 Options/Peripherals/_General..., Item
+Options/Peripherals/_Sound..., Item
 Options/Peripherals/_Disk..., Item
 
 Options/_RZX..., Item
