+++ peripherals/sound/covox.c
