+++ peripherals/sound/covox.h
