+++ peripherals/sound/specdrum.h
